Time: 1489 ms (5.10%), Space: 265.3 MB (5.10%) - LeetHub

+++ 2008-maximum-earnings-from-taxi/2008-maximum-earnings-from-taxi.cpp
@@ -2,7 +2,7 @@
 class Solution {
 public:
     long long maxTaxiEarnings(int n, vector<vector<int>>& rides) {
-        vector<long long> dp(n+1, 0);
+        vector<long> dp(n+1, 0);
         unordered_map<int, vector<pair<int,long>>> map;
         for(auto i: rides){
             map[i[1]].push_back({i[0], i[1]-i[0]+i[2]});
